Close #58: Implement non session based CAPTCHAs
The basic idea is that it should be sufficient to have a secret key on
the server, and to have input fields for the CAPTCHA code and its MAC.
Furthermore, attackers should not be able to find such pairs without
knowing the secret key.  An HMAC-SHA256 generated from the code and a
properly initialized `random` configuration option as key is supposed
to fulfill these requirements.