Put configuration into a meta element instead of a script
While the latter is not bad, the former is slightly better decoupled.
We use single quotes to enclose the attribute value and a suitable set
of JSON flags, to produce the minimal amount of JSON while still being
safe.

In the long run, we'll want to include all the JS stuff only when
really necessary, though, which may give the opportunity to put the
configuration into a data attribute.