Close #58: Implement non session based CAPTCHAs
The basic idea is that it should be sufficient to have a secret key on
the server, and to have input fields for the CAPTCHA code and its MAC.
Furthermore, attackers should not be able to find such pairs without
knowing the secret key.  An HMAC-SHA256 generated from the code and a
properly initialized `random` configuration option as key is supposed
to fulfill these requirements.

However, once a suitable combination of code and its MAC have been
found, arbitrary replay attacks would be possible.  We restrict that
by adding the timestamp of the code generation into the mix, and let
that code expire after five minutes.  That is inferior to a nonce, but
a nonce would require state, what we want to avoid, and after all, the
text CAPTCHA could easily be parsed by a bot anyway.  So we leave that
for more advanced solutions like Cryptographp_XH.

In the hopefully unlikely case that spammers know the secret key, they
could compute rainbow tables for arbitrary timestamps (after all, there
are only 100,000 code combinations).  We mitigate that by adding a four
byte random salt.