Close #83: Support direct SMTP mail
We take Olaf's modifications[1] as starting point, and apply some
further modifications on top.  Most notably we do not add the config
option `smtp_email_address_from` because users can and should use the
respective form definition fields for this purpose.  Furthermore, we
ensure that SMTP debug output is only ever shown in the back-end, and
add a system check regarding access protection of config.php.

We also document everything, hopefully understandable for end-users.

Furthermore, we patch PHPMailer to suppress the link to the Wiki page
in case of SMTP connection failures, since it is unlikely to be helpful
for Advancedform_XH users.

[1] <https://cmsimpleforum.com/viewtopic.php?f=12&t=3451&p=87686#p87686>

Co-authored-by: Olaf Penschke <35469568+olape-git@users.noreply.github.com>